Added code for change progress on seek bar

+++ app/src/main/java/com/example/photoeditor/Interface/EditImageFragmentListener.java
@@ -2,8 +2,8 @@ package com.example.photoeditor.Interface;
 
 public interface EditImageFragmentListener {
     void onBrightnessChanged(int brightness);
-    void onSaturationChanged(int saturation);
-    void onContrastChanged(int contrast);
+    void onSaturationChanged(float saturation);
+    void onContrastChanged(float contrast);
     void onEditStarted();
     void onEditCompleted();
 }
